+++ app/src/main/java/com/wasingun/seller_lounge/data/datasource/PostDataSource.kt
@@ -1,14 +1,14 @@
 package com.wasingun.seller_lounge.data.datasource
 
-import com.wasingun.seller_lounge.data.enums.ProductCategory
-import com.wasingun.seller_lounge.data.model.DocumentContent
-import com.wasingun.seller_lounge.data.model.ImageContent
+import com.wasingun.seller_lounge.data.model.ProductCategory
+import com.wasingun.seller_lounge.data.model.attachedcontent.DocumentContent
+import com.wasingun.seller_lounge.data.model.attachedcontent.ImageContent
 import com.wasingun.seller_lounge.data.model.post.PostInfo
 import com.wasingun.seller_lounge.data.model.post.UserInfo
 import com.wasingun.seller_lounge.network.ApiResponse
 import kotlinx.coroutines.flow.Flow
 
-interface RemotePostDataSource {
+interface PostDataSource {
 
     fun getPostList(
         onComplete: () -> Unit,
